+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 ## Unreleased
 
+- Allow additional properties when creating objects;
+- Ability to create objects by only providing `:object-type`;
+- Update dependencies, including BACnet4J 6.0.0.
+
 ## [1.1.9] - 2021-01-22
 - Fix bug where a nil `:device-id` could be returned by `local-device-backup`.